move sed methods to sed.py
